Resolve conflicts from the consensus/ledger refactor vs main:
- rcl/engine.go: adopt main's IsValidating() helper (terse comment) and
  route GetJSON's validating flag through it.
- service.go: keep the refactor's split (moved funcs live in lifecycle/cache);
  preserve main's auto-merged #1105 PrewarmSignature in SubmitOpenLedgerTx.
- lifecycle.go: port main's #1091 flag-ledger NegativeUNL transition
  (applyFlagLedgerNegativeUNL + 3 close-path call sites) into the new home.

+++ codec/binarycodec/codec.go
@@ -188,7 +188,7 @@ func EncodeForSigningBatch(json map[string]any) (string, error) {
 
 	txIDsLengthType := &types.UInt32{}
 	txIDsLength := len(txIDsInterface)
-	if txIDsLength > math.MaxUint32 {
+	if uint64(txIDsLength) > math.MaxUint32 {
 		return "", ErrBatchTxIDsLengthTooLong
 	}
 	txIDsLengthBytes, err := txIDsLengthType.FromJSON(uint32(txIDsLength))
